% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_plot.R
\name{aggregate_plot}
\alias{aggregate_plot}
\title{Aggregate and Visualize Selection Probabilities}
\usage{
aggregate_plot(
  nburnin,
  niter,
  result,
  taxa_table,
  level = "Phylum",
  label_size = 10
)
}
\arguments{
\item{nburnin}{Integer, number of burn-in iterations to skip in the analysis.}

\item{niter}{Integer, number of iterations considered for aggregating results.}

\item{result}{List, containing results from a Gibbs sampling algorithm, specifically \code{betahat}
for estimated coefficients and \code{gammaresult} for variable inclusion results.}

\item{taxa_table}{Data frame, containing taxonomic information for each taxon. Must include
a column corresponding to \code{level} that contains taxonomic classifications.}

\item{level}{Character, specifying the taxonomic level to aggregate and visualize data.
Defaults to "Phylum".}

\item{label_size}{Numeric, size of the label text in the plot, default is 10.}
}
\value{
A ggplot object displaying the aggregated selection probabilities as a bar plot,
grouped by the specified taxonomic level. Each group is annotated with the number of
taxa included in that group.
}
\description{
This function aggregates results from a Gibbs sampling process and visualizes selection
probabilities of taxa at specified taxonomic levels, comparing positive and negative selections.
It calculates the counts of negative and positive selections for each taxon, merges these with
taxonomic data, and produces a bar plot grouped by the specified taxonomic level.
}
\examples{
\dontrun{
  result <- list(betahat = matrix(rnorm(2000), 100, 20),
                 gammaresult = sample(0:1, 100, replace = TRUE))
  taxa_table <- data.frame(Phylum = rep(c("Firmicutes", "Bacteroidetes"), each = 50))
  plot <- aggregate_plot(nburnin = 500, niter = 1500, result = result,
                         taxa_table = taxa_table, level = "Phylum")
  print(plot)
}
}
