% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tax.R
\name{create_tax}
\alias{create_tax}
\title{Export Cleaned Taxonomic Paths to a File}
\usage{
create_tax(ps, tree_file)
}
\arguments{
\item{ps}{A \code{phyloseq} object containing taxonomic data.}

\item{tree_file}{Character, the filepath where the cleaned taxonomic paths will be written.}
}
\value{
None; the function writes the cleaned taxonomic paths directly to the specified file.
}
\description{
This function processes taxonomic data from a \code{phyloseq} object, cleans it by removing unwanted
characters and filtering out non-informative entries, and writes the resulting taxonomic paths
to a specified file. Each path is a concatenation of the taxonomic levels for each organism,
providing a clear hierarchical view of the taxonomy.
}
\details{
The function verifies that the taxonomic data are structured with taxa as rows. It then converts
the taxonomic table to a dataframe and cleanses it by removing periods and ignoring entries labeled
as "unknown" or NA. The cleaned taxonomic names are then concatenated to form paths, which are
written to the provided file, each path on a new line. This processing is essential for ensuring
the clarity and usability of taxonomic data in downstream analyses or visualizations.
}
\examples{
\dontrun{
  library(phyloseq)
  data(GlobalPatterns)
  create_tax(GlobalPatterns, "path/to/your/clean_taxonomic_paths.txt")
}

}
