% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tax_annot.R
\name{create_tax_annot}
\alias{create_tax_annot}
\title{Create Taxonomic Annotations for Visualization}
\usage{
create_tax_annot(ps, annotation_file, level = "Genus")
}
\arguments{
\item{ps}{A \code{phyloseq} object containing microbiome or other taxonomic data.}

\item{annotation_file}{Character, path to the file where annotations will be saved.}

\item{level}{Character, the taxonomic level at which annotations are to be generated.
Default is "Genus". Supported levels depend on the taxonomy table in the \code{phyloseq} object.}
}
\value{
None; this function writes directly to a file specified by \code{annotation_file}.
}
\description{
This function generates annotation files from a phyloseq object based on specified taxonomic levels.
It validates the taxonomic structure, computes relative abundances, and assigns colors to different
taxonomic levels for visualization purposes. The annotations are saved to a file which can be used in
various tree visualization tools to enhance the presentation of phylogenetic data.
}
\details{
The function first checks if taxa are correctly set as rows in the \code{phyloseq} object. It then extracts
the OTU table and taxonomy table, identifies the unique taxa at the specified level, and assigns a
unique color to each. It calculates the relative abundance of each taxon, scales these values, and
writes detailed annotation data into the specified file. This file includes settings for colors, sizes,
and other visual attributes for each taxon, facilitating detailed customization of phylogenetic tree visualizations.
}
\examples{
\dontrun{
  library(phyloseq)
  data(GlobalPatterns)
  create_tax_annot(GlobalPatterns, "path/to/your/annotation_file.txt", level="Genus")
}

}
