% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tax_selected.R
\name{create_tax_selected}
\alias{create_tax_selected}
\title{Generate Detailed Taxonomic Annotations for Selected Taxa}
\usage{
create_tax_selected(
  ps,
  nburnin,
  niter,
  result,
  annotation_file,
  level = "Genus"
)
}
\arguments{
\item{ps}{A \code{phyloseq} object containing taxonomic data.}

\item{nburnin}{Integer, number of burn-in iterations to disregard in the analysis.}

\item{niter}{Integer, number of iterations considered for analysis after burn-in.}

\item{result}{List, containing statistical results including 'betahat' and 'gammaresult'.}

\item{annotation_file}{Character, the file path where the annotations will be written.}

\item{level}{Character, taxonomic level at which annotations are generated (default "Genus").}
}
\value{
None; this function writes annotations directly to the specified file.
}
\description{
This function processes a \code{phyloseq} object to create annotations for taxa that are notably represented
in the results of variable selection model. It extracts taxonomic and abundance data, evaluates the
significance of each taxa's representation in the results, and writes annotations with specific visual
attributes to a file.
}
\details{
The function starts by ensuring the \code{phyloseq} object is properly formatted with taxa as rows.
It calculates positive and negative counts for each taxon's representation in the results and
selects taxa based on a threshold. Each selected taxon's annotation includes visual attributes
like color and shape, tailored based on its statistical representation. These annotations are
designed to enhance visualizations, particularly phylogenetic trees, by providing clear indications
of the taxa's significance in the study.

The function also checks for taxonomic rank format correctness and handles unknown values gracefully,
ensuring that the output file contains only well-formed and meaningful data. The colors and shapes
used in the annotations can be adjusted within the function to match specific visualization needs.
}
