% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detailed_tax.R
\name{detailed_tax}
\alias{detailed_tax}
\title{Detailed Taxonomic Tree Plot with Highlighted Nodes}
\usage{
detailed_tax(
  ps,
  anno.data,
  alpha = 0.2,
  anno.depth = 3,
  anno.x = 10,
  anno.y = 40,
  layout = "circular"
)
}
\arguments{
\item{ps}{A phyloseq object containing the microbiome data.}

\item{anno.data}{A data frame containing annotation data with columns for nodes and colors.}

\item{alpha}{A numeric value specifying the transparency level for highlighting nodes. Default is 0.2.}

\item{anno.depth}{An integer specifying the depth of annotation. Default is 3.}

\item{anno.x}{An integer specifying the x-coordinate offset for annotations. Default is 10.}

\item{anno.y}{An integer specifying the y-coordinate offset for annotations. Default is 40.}

\item{layout}{A character string specifying the layout of the tree plot. Default is "circular".}
}
\value{
A ggplot object representing the annotated taxonomic tree.
}
\description{
This function generates a detailed taxonomic tree plot with highlighted nodes and customized labels.
The function takes a phyloseq object and annotation data to create a tree plot, highlighting selected nodes,
and annotating them with specified labels.
}
\examples{
\dontrun{
ps <- phyloseq_object  # Replace with actual phyloseq object
anno.data <- data.frame(node = c("node1", "node2"), color = c("red", "blue"))  # Replace with actual annotation data
plot <- detailed_tax(ps, anno.data, alpha=0.2, anno.depth=3, layout="rectangular")
print(plot)
}

}
