% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_plot.R
\name{effect_plot}
\alias{effect_plot}
\title{Plot Effect Sizes from Bayesian Model Results}
\usage{
effect_plot(result, ps, nburnin, niter, mode = "mean", level = "Genus")
}
\arguments{
\item{result}{A list from Bayesian variable selection containing at least the betahat
coefficients and the gamma result selection frequencies.}

\item{ps}{A \code{phyloseq} object containing microbiome data which will be aggregated
at the specified taxonomic level.}

\item{nburnin}{The number of burn-in iterations in the Bayesian model; these initial iterations
will be excluded from the effect size calculation.}

\item{niter}{The total number of iterations after burn-in, used to determine the threshold
for significant selection in the visualization.}

\item{mode}{A character string specifying the method to summarize the effect sizes across iterations.
Options are "mean" or "median". Defaults to "mean".}

\item{level}{The taxonomic level at which to aggregate data in the \code{phyloseq} object.
Defaults to "Genus".}
}
\value{
A ggplot object representing the effect size plot, with taxa on the y-axis and
effect sizes on the x-axis. Bars are filled according to the sign of the effect size.
}
\description{
This function calculates and plots the effect sizes derived from Bayesian variable
selection model results. It processes the results to compute effect sizes based on
specified summary statistics (mean or median) after aggregating data at a given taxonomic level.
The effect sizes are then visualized in a bar plot with bars colored based on their
sign (positive or negative).
}
