% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_sign.R
\name{effect_sign}
\alias{effect_sign}
\title{Generate Effect Significance for Taxonomic Levels}
\usage{
effect_sign(result, ps, nburnin, niter, mode = "mean", level = "Genus")
}
\arguments{
\item{result}{A list containing the \code{betahat} matrix and \code{gammaresult} vector from the model results.}

\item{ps}{A phyloseq object containing the microbiome data.}

\item{nburnin}{An integer specifying the number of burn-in iterations.}

\item{niter}{An integer specifying the total number of iterations.}

\item{mode}{A character string specifying the mode of calculation (currently unused, reserved for future use). Default is "mean".}

\item{level}{A character string specifying the taxonomic level at which to aggregate the data (e.g., "Genus"). If NULL, no aggregation is performed.}
}
\value{
A data frame containing taxa names and their corresponding effect significance ("positive" or "negative").
}
\description{
This function computes the effect size and significance (positive or negative) for taxa at a specified taxonomic level based on the provided phyloseq object and model results.
}
\details{
This function first aggregates the taxa in the phyloseq object to the specified taxonomic level (if provided). It then ensures the dimensions of the taxonomy table align with the dimensions of the \code{betahat} matrix from the model results. The function combines the taxa names with the selection results (\code{gammaresult}), filters taxa based on a threshold, calculates the effect size for each taxon, and determines the effect's significance (positive or negative). Finally, it returns a data frame with taxa names and their effect significance, excluding taxa labeled as "unknown" or "uncultured".
}
\examples{
# Assuming 'result' is a list with 'betahat' and 'gammaresult', 'ps' is a phyloseq object
# effect_sign(result, ps, nburnin=100, niter=1000, level="Genus")
}
