% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_size.R
\name{effect_size}
\alias{effect_size}
\title{Calculate Effect Sizes for Bayesian Variable Selection Model Results}
\usage{
effect_size(result, ps, nburnin, niter, mode = "mean", level = "Genus")
}
\arguments{
\item{result}{A list containing the results from a Bayesian variable selection model,
expected to include elements \code{betahat} and \code{gammaresult} where \code{betahat} are the
coefficient estimates across iterations and \code{gammaresult} indicates selection frequency.}

\item{ps}{A \code{phyloseq} object containing microbiome data, which will be aggregated at the specified
taxonomic level before analysis.}

\item{nburnin}{Integer specifying the number of burn-in iterations in the model, used to
exclude these iterations from effect size calculations.}

\item{niter}{Integer specifying the total number of iterations after burn-in, used to determine
the threshold for considering a variable significantly selected.}

\item{mode}{Character string specifying how to summarize the effect sizes across iterations.
Options are "mean" or "median". Defaults to "mean".}

\item{level}{The taxonomic level at which to aggregate data in the \code{phyloseq} object.
Defaults to "Genus".}
}
\value{
Returns a data frame with taxa and their corresponding effect sizes, ordered
by effect size in decreasing order.
}
\description{
This function calculates effect sizes for variables in a Bayesian variable selection model
based on the \code{betahat} coefficients and selection frequency across iterations. The function
allows users to choose between mean or median to summarize the effect sizes and requires
that the number of taxa in the \code{phyloseq} object matches the number of \code{betahat} coefficients.
}
