% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_selected_id}
\alias{get_selected_id}
\title{Get Selected Taxonomic IDs}
\usage{
get_selected_id(ps, anno.data)
}
\arguments{
\item{ps}{A phyloseq object containing the microbiome data.}

\item{anno.data}{A data frame containing annotation data with columns for nodes.}
}
\value{
A vector of selected taxonomic IDs.
}
\description{
This function filters the taxonomic table of a phyloseq object to retain only the taxa that match
the specified nodes in the annotation data.
}
\examples{
\dontrun{
ps <- phyloseq_object  # Replace with actual phyloseq object
anno.data <- data.frame(node = c("g__Bacteroides", "f__Lachnospiraceae"), color = c("red", "blue"))  # Replace with actual annotation data
selected_ids <- get_selected_id(ps, anno.data)
print(selected_ids)
}

}
