% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbsgamma.R
\name{gibbsgamma}
\alias{gibbsgamma}
\title{Gibbs Sampling for Bayesian Variable Selection}
\usage{
gibbsgamma(
  nburnin,
  niter,
  p,
  nop,
  Y,
  X,
  N,
  a,
  Q,
  n,
  tau,
  nu,
  omega,
  seed,
  predict,
  stand,
  display,
  split_rate = 0.7
)
}
\arguments{
\item{nburnin}{Integer, number of burn-in iterations for the Gibbs sampler.}

\item{niter}{Integer, number of iterations after burn-in.}

\item{p}{Integer, number of predictors.}

\item{nop}{Integer, initial number of predictors to include.}

\item{Y}{Numeric vector or matrix, response variable(s).}

\item{X}{Numeric matrix, predictor variables.}

\item{N}{Numeric matrix, normalization matrix for predictors.}

\item{a}{Numeric vector, hyperparameters for the gamma prior (usually negative number).}

\item{Q}{Numeric matrix, phylogenetic matrix.}

\item{n}{Integer, number of observations in Y and rows in X.}

\item{tau}{Numeric, hyperparameter scaling the precision matrix N.}

\item{nu}{Numeric, degrees of freedom for the inverse gamma distribution of sigma^2.}

\item{omega}{Numeric, scale parameter for the inverse gamma distribution of sigma^2.}

\item{seed}{Integer, seed for random number generation to ensure reproducibility.}

\item{predict}{Logical, whether to perform prediction on the test set.}

\item{stand}{List, contains standardized parameters including 'mux' and 'Sx' for X, and 'muy' and 'Sy' for Y.}

\item{display}{Logical, whether to display progress of the Gibbs sampler.}

\item{split_rate}{Numeric, proportion of the data to be used as the training set (default is 0.7).}
}
\value{
A list containing:
\itemize{
\item \code{gamma}: Matrix of inclusion indicators for variables across all iterations.
\item \code{betahat}: Matrix of estimated coefficients.
\item \code{gammaresult}: Sum of gamma indicators across post-burn-in iterations.
\item \code{MSE}: Mean squared error of predictions on the testing set.
\item \code{nselect}: Number of selected variables in each iteration.
\item \code{Yhat}: Predicted values for the testing set.
}
}
\description{
This function implements a Gibbs sampling algorithm for variable selection in a Bayesian
framework, using a gamma-shrinkage prior. It splits the data into training and testing sets,
performs the sampling, and optionally predicts responses based on selected models.
}
