% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{phy_to_tax}
\alias{phy_to_tax}
\title{Convert Phyloseq Object to Phylogenetic Tree Data}
\usage{
phy_to_tax(
  physeq,
  use_abundance = TRUE,
  node.size.scale = 1,
  node.size.offset = 1
)
}
\arguments{
\item{physeq}{A phyloseq object containing the microbial community data.}

\item{use_abundance}{A logical value indicating whether to use abundance data. Default is TRUE.}

\item{node.size.scale}{A numeric value specifying the scaling factor for node sizes based on relative abundance. Default is 1.}

\item{node.size.offset}{A numeric value specifying the offset for node sizes. Default is 1.}
}
\value{
Returns a \code{treedata} object that includes the phylogenetic tree and associated data such as node sizes and classifications.
}
\description{
The \code{phy_to_tax} function converts a phyloseq object into a phylogenetic tree data structure suitable for use with various phylogenetic analysis and visualization tools.
}
\details{
The function processes the phyloseq object to ensure that taxa are rows, fixes any duplicate taxonomy, and summarizes the taxa at each level. It then constructs a tree data structure using the \code{createtree} function.
}
\examples{
\dontrun{
# Load necessary libraries
library(phyloseq)
library(ggtree)

# Example usage
data(GlobalPatterns)
tree_data <- phy_to_tax(GlobalPatterns)
}

}
