% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottax.R
\name{plottax}
\alias{plottax}
\title{Visualize Taxonomic Tree with Annotations}
\usage{
plottax(tree, anno.data, alpha = 0.2, anno.depth = 3, anno.x = 10, anno.y = 40)
}
\arguments{
\item{tree}{A phylogenetic tree object compatible with \code{ggtree}, representing the evolutionary relationships among the taxa.}

\item{anno.data}{A data frame containing annotation information for the tree. It should include columns for \code{node} and \code{color}.}

\item{alpha}{A numeric value specifying the transparency level for the highlighted clades. Default is 0.2.}

\item{anno.depth}{A numeric value specifying the depth of annotation. Only nodes at or below this depth will be annotated with short labels. Default is 3.}

\item{anno.x}{A numeric value specifying the x-coordinate for annotation placement. Default is 10.}

\item{anno.y}{A numeric value specifying the y-coordinate for annotation placement. Default is 40.}
}
\value{
Returns a ggtree plot object with highlighted clades and labels based on the specified taxonomic annotations.
}
\description{
The \code{plottax} function generates a circular phylogenetic tree visualization with clades highlighted and labeled based on specified taxonomic annotations.
}
\details{
The function proceeds through several steps to generate the final plot:
\itemize{
\item Generates the initial ggtree plot.
\item Highlights clades based on the provided annotation data.
\item Adds labels to the clades based on their taxonomic level.
\item Adds a legend for the annotations.
}
}
\examples{
\dontrun{
# Load necessary libraries
library(ggtree)
library(dplyr)

# Example usage
tree <- read.tree("path/to/tree_file") # Load your tree
anno.data <- data.frame(node = c("node1", "node2"), color = c("red", "blue"))
p <- plottax(tree, anno.data, alpha = 0.2, anno.depth = 3, anno.x = 10, anno.y = 40)
print(p)
}

}
