% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{result}
\alias{result}
\title{Bayesian-tree based Model Fitting for Microbiome Data}
\usage{
result(ps, nburnin, niter, a, level = NULL, response)
}
\arguments{
\item{ps}{A \code{phyloseq} object containing microbiome data.}

\item{nburnin}{Integer, number of burn-in iterations for the Gibbs sampler.}

\item{niter}{Integer, number of sampling iterations after burn-in.}

\item{a}{Numeric vector or scalar, hyperparameters for the model (choose it carefully!).}

\item{level}{Optional character, taxonomic level at which to aggregate data (e.g., "Phylum").
If NULL, uses the original taxonomy without aggregation.}

\item{response}{Character, name of the response variable in the sample data.}
}
\value{
Returns a list containing the results from the Gibbs sampler, including parameter estimates
and diagnostics from the sampling process.
}
\description{
This function processes microbiome data, aggregates it at specified taxonomic levels,
and fits a Bayesian-tree based model. It checks the integrity of the
operational taxonomic unit (OTU) table, scales the data, prepares the phylogenetic matrix,
and runs a Gibbs sampling algorithm to estimate model parameters.
}
\details{
The function first validates the OTU table structure, ensuring that taxa are properly represented.
It then optionally aggregates the OTU data to a specified taxonomic level. Response variables and
predictors are extracted from the \code{phyloseq} object. Data is then log-transformed and standardized.
A phylogenetic variance-covariance matrix is computed and adjusted. The function finally invokes the
\code{gibbsgamma} function for fitting the Bayesian model.
}
