% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_rel.R
\name{tax_rel}
\alias{tax_rel}
\title{Visualize Phylogenetic Tree with Taxonomic Annotations and Relative Abundance Bars}
\usage{
tax_rel(
  tree,
  anno.data,
  alpha = 0.2,
  anno.depth = 3,
  anno.x = 10,
  anno.y = 40,
  scale_size = 0.5
)
}
\arguments{
\item{tree}{A phylogenetic tree object compatible with \code{ggtree}, representing the evolutionary relationships among the taxa.}

\item{anno.data}{A data frame containing annotation information for the tree. It should include columns for \code{node} and \code{color}.}

\item{alpha}{A numeric value specifying the transparency level for the highlighted clades. Default is 0.2.}

\item{anno.depth}{A numeric value specifying the depth of annotation. Only nodes at or below this depth will be annotated with short labels. Default is 3.}

\item{anno.x}{A numeric value specifying the x-coordinate for annotation placement. Default is 10.}

\item{anno.y}{A numeric value specifying the y-coordinate for annotation placement. Default is 40.}

\item{scale_size}{A numeric value to scale the size of the relative abundance bars and tree plot. Default is 0.5.}
}
\value{
Returns a ggtree plot object with highlighted clades, labels, and relative abundance bars based on the specified taxonomic annotations.
}
\description{
The \code{tax_rel} function generates a circular phylogenetic tree visualization with clades highlighted, annotated based on specified taxonomic levels, and includes bars representing relative abundance outside the tree.
}
\details{
The function proceeds through several steps to generate the final plot:
\itemize{
\item Generates the initial ggtree plot.
\item Highlights clades based on the provided annotation data.
\item Propagates phylum information from internal nodes to tips.
\item Adds labels to the clades based on their taxonomic level.
\item Adds bars representing relative abundance outside the tree.
\item Adds a legend for the annotations.
}
}
\examples{
\dontrun{
# Load necessary libraries
library(ggtree)
library(dplyr)

# Example usage
tree <- read.tree("path/to/tree_file") # Load your tree
anno.data <- data.frame(node = c("node1", "node2"), color = c("red", "blue"))
p <- tax_rel(tree, anno.data, alpha = 0.2, anno.depth = 3, anno.x = 10, anno.y = 40, scale_size = 0.5)
print(p)
}

}
