% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxview.R
\name{taxview}
\alias{taxview}
\title{Visualize Phylogenetic Tree with Taxonomic Clade Highlights}
\usage{
taxview(
  ps,
  tree,
  branch_thickness = 0.5,
  layout = "circular",
  level = "Phylum"
)
}
\arguments{
\item{ps}{A phyloseq object containing the microbial community data, including the OTU table and taxonomy table.}

\item{tree}{A phylogenetic tree object compatible with \code{ggtree}, representing the evolutionary relationships among the taxa in the phyloseq object.}

\item{branch_thickness}{A numeric value specifying the thickness of the branches in the phylogenetic tree. Default is 0.5.}

\item{layout}{A character string specifying the layout of the phylogenetic tree. Default is 'circular'. Other options include 'rectangular', 'slanted', etc.}

\item{level}{A character string specifying the taxonomic level to highlight in the tree (e.g., "Phylum", "Genus"). Default is "Phylum".}
}
\value{
Returns a ggtree plot object with highlighted clades and labels based on the specified taxonomic level.
}
\description{
The \code{taxview} function generates a phylogenetic tree visualization with clades highlighted and labeled based on a specified taxonomic level from a phyloseq object.
This function facilitates the interpretation of microbial community data by incorporating taxonomic information directly into the tree plot.
}
\details{
The function proceeds through several steps to generate the final plot:
\itemize{
\item Extracts the taxonomy table from the phyloseq object.
\item Generates a color palette for the unique taxonomic groups at the specified level.
\item Maps the tips of the phylogenetic tree to their corresponding taxonomic groups.
\item Creates the initial ggtree plot with the specified branch thickness and layout.
\item Highlights and labels clades for each unique taxonomic group:
\itemize{
\item Identifies the internal node label in the tree corresponding to each taxonomic group.
\item Highlights the clade associated with that node using the assigned color.
\item Adds clade labels outside the highlighted clades.
}
\item Adds a legend mapping the taxonomic groups to their colors.
}
}
\examples{
\dontrun{
# Load necessary libraries
library(phyloseq)
library(ggtree)
library(scales)

# Assuming `ps1` is your phyloseq object and `tree` is the corresponding phylogenetic tree
p <- taxview(ps1, tree, branch_thickness = 0.5, layout = 'circular', level = "Phylum")

# Display the plot
print(p)
}

}
